Add "docs" build and CI step
Adds a `zig build docs` step to build documentation website,
and a new GitHub Workflow that publishes the website to GitHub Pages.

This workflow is divided into two jobs:

- build: builds the documentation and uploads it
- publish: downloads the documentation and publishes it

These are separate jobs to minimize permissions available
to the build job.

To use this, you have to take the following steps:

1. Go to **Settings** for the repository
2. Select **Pages** on the left under *Code and automation*
3. Under *Build and deployment* set **Source** to **GitHub Actions**

Note that this workflow dispatches on two events:

- after every push to master
- `workflow_dispatch`: this allows manually running the workflow
  from its *Actions* page if something went wrong

---

One of the points made by @joachimschmidt557 in #13 was that
the generated docs must explicitly call out that the documentation is
unstable/WIP.

Because this uses the same tooling as [Zig std][1],
the generated pages include the following header at the top:

> This is a beta autodoc build; expect bugs and missing information.

  [1]: https://ziglang.org/documentation/master/std

Resolves #13